Move events module to watch
