+++ tests/test_events.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python3 -m pytest
-from plextraktsync.events import ActivityNotification
 from plextraktsync.watch.EventDispatcher import EventDispatcher
 from plextraktsync.watch.EventFactory import EventFactory
+from plextraktsync.watch.events import ActivityNotification
 from tests.conftest import load_mock
